third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristics/characteristics-found-with-uuid.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
    return getTwoMeasurementIntervalCharacteristicsService()
      .then(({service}) => Promise.all([
        service.getCharacteristics(measurement_interval.alias),
        service.getCharacteristics(measurement_interval.name),
        service.getCharacteristics(measurement_interval.uuid)]))
    .then(characteristics_arrays => {
      characteristics_arrays.forEach(characteristics => {
        assert_equals(characteristics.length, 2);
        assert_equals(characteristics[0].uuid,
                      measurement_interval.uuid);
        assert_equals(characteristics[1].uuid,
                      measurement_interval.uuid);
      });
    });
}, 'Find characteristics with UUID in service.');
</script>
